refactor: move Prism comp to default exp
