+++ packages/astro-prism/Prism.astro
@@ -1,6 +1,6 @@
 ---
 import Prism from 'prismjs';
-import { addAstro } from '../index.mjs';
+import { addAstro } from './internal.mjs';
 import loadLanguages from 'prismjs/components/index.js';
 
 export interface Props {
