+++ packages/astro-prism/index.js
@@ -1,0 +1,1 @@
+export { default as Prism } from './Prism.astro';
