+++ packages/astro-prism/internal.d.ts
